move check version
